feat: rename auto set properties to default permissions
