+++ apps/client/src/components/admin/manage/users/modals/manage-permissions-modal.tsx
@@ -3,73 +3,20 @@ import { Modal } from "components/modal/Modal";
 import { useModal } from "state/modalState";
 import { useTranslations } from "next-intl";
 import { ModalIds } from "types/ModalIds";
-import {
-  getPermissions,
-  defaultPermissions,
-  PermissionNames,
-  Permissions,
-} from "@snailycad/permissions";
+import { getPermissions, PermissionNames, Permissions } from "@snailycad/permissions";
 import { FormField } from "components/form/FormField";
 import { Toggle } from "components/form/Toggle";
 import { Form, Formik } from "formik";
 import useFetch from "lib/useFetch";
 import { Loader, Button, TextField } from "@snailycad/ui";
 import type { GetManageUserByIdData, PutManageUserPermissionsByIdData } from "@snailycad/types/api";
+import { usePermissionsModal } from "hooks/use-permissions-modal";
 
 interface Props {
   user: Pick<GetManageUserByIdData, "permissions" | "id" | "roles" | "rank">;
   isReadOnly?: boolean;
   onUpdate?(user: PutManageUserPermissionsByIdData): void;
 }
-
-const DEPRECATED_PERMISSIONS = [
-  Permissions.ViewDLExams,
-  Permissions.ManageDLExams,
-  Permissions.ViewWeaponExams,
-  Permissions.ManageWeaponExams,
-];
-
-const filteredAdminPermissions = defaultPermissions.allDefaultAdminPermissions.filter(
-  (p) => !defaultPermissions.defaultCourthousePermissions.includes(p),
-);
-
-const groups = [
-  {
-    name: "Admin",
-    permissions: filteredAdminPermissions,
-  },
-  {
-    name: "Courthouse (Admin)",
-    permissions: defaultPermissions.defaultCourthousePermissions,
-  },
-  {
-    name: "LEO",
-    permissions: defaultPermissions.defaultLeoPermissions,
-  },
-  {
-    name: "Dispatch",
-    permissions: defaultPermissions.defaultDispatchPermissions,
-  },
-  {
-    name: "EMS/FD",
-    permissions: defaultPermissions.defaultEmsFdPermissions,
-  },
-  {
-    name: "Citizen related",
-    permissions: [
-      ...defaultPermissions.defaultTowPermissions,
-      ...defaultPermissions.defaultTaxiPermissions,
-    ],
-  },
-  {
-    name: "Other",
-    permissions: defaultPermissions.otherDefaultPermissions,
-  },
-  {
-    name: "Owner",
-    permissions: [Permissions.ManageCADSettings],
-  },
-];
 
 export function ManagePermissionsModal({ user, onUpdate, isReadOnly }: Props) {
   const [search, setSearch] = React.useState("");
@@ -79,6 +26,7 @@ export function ManagePermissionsModal({ user, onUpdate, isReadOnly }: Props) {
   const { closeModal, isOpen } = useModal();
   const userPermissions = getPermissions(user);
   const { state, execute } = useFetch();
+  const { DEPRECATED_PERMISSIONS, groups, handleToggleAll } = usePermissionsModal({ isReadOnly });
 
   async function onSubmit(data: typeof INITIAL_VALUES) {
     if (isReadOnly) return;
@@ -92,38 +40,6 @@ export function ManagePermissionsModal({ user, onUpdate, isReadOnly }: Props) {
     if (json.id) {
       closeModal(ModalIds.ManagePermissions);
       onUpdate?.(json);
-    }
-  }
-
-  function handleToggleAll(
-    group: (typeof groups)[number],
-    values: Record<Permissions, boolean>,
-    setValues: any,
-  ) {
-    if (isReadOnly) return;
-
-    const groupPermissionValues = Object.entries(values).filter(([permission]) => {
-      return group.permissions.includes(permission as Permissions);
-    });
-    const areAllChecked = groupPermissionValues.every(([, b]) => b);
-
-    if (areAllChecked) {
-      const obj = groupPermissionValues.reduce(
-        (ac, [permission]) => ({ ...ac, [permission]: false }),
-        {},
-      );
-
-      setValues({ ...values, ...obj });
-    } else {
-      const obj = group.permissions.reduce(
-        (ac, cv) => ({
-          ...ac,
-          [cv]: true,
-        }),
-        {},
-      );
-
-      setValues({ ...values, ...obj });
     }
   }
